backend/requirements.txt
# Web framework
fastapi==0.110.0
uvicorn[standard]==0.27.1

# Database
sqlalchemy==2.0.25
# Uwaga: 'sqlalchemy.orm' jest częścią pakietu 'sqlalchemy' i nie trzeba go instalować osobno.
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Configuration & Environment
python-dotenv==1.0.0
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic[email]

# Cloud Storage
boto3==1.34.48
azure-storage-blob==12.19.0

# Other utilities
python-multipart==0.0.9

# Testing
pytest==8.0.2
httpx==0.26.0
pytest-sugar
